backend/requirements.txt
Flask==3.1.2
flask-cors==6.0.2
spacy==3.8.11
langdetect==1.0.9
rapidfuzz==3.10.1
opencv-python-headless==4.12.0.88
pytesseract==0.3.13
Pillow==11.1.0
deep-translator==1.11.4
python-dotenv==1.0.1